app/requirements.txt
numpy
kafka-python
confluent-kafka
redshift_connector
pandas
multiprocess